requirements.txt
asgiref==3.8.1
boto3==1.34.14
botocore==1.34.162
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
crispy-bootstrap5==0.7
cryptography==43.0.3
defusedxml==0.7.1
Django==5.0.1
django-allauth==0.59.0
django-cleanup==8.0.0
django-crispy-forms==2.1
django-debug-toolbar==4.2.0
django-filter==23.5
django-guardian==2.4.0
django-ninja==0.17.0
django-storages==1.14.2
djangorestframework==3.15.2
idna==3.10
iniconfig==2.0.0
jmespath==1.0.1
oauthlib==3.2.2
packaging==24.1
pillow==10.2.0
pluggy==1.5.0
pycparser==2.22
pydantic==1.10.18
PyJWT==2.9.0
pytest==7.4.4
pytest-django==4.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-magic==0.4.27
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
s3transfer==0.10.3
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
